+++ examples/window-app/common/src/WindowApp.cpp
@@ -44,7 +44,7 @@ void OnTriggerEffect(Identify * identify)
     if (identify->mCurrentEffectIdentifier == Clusters::Identify::EffectIdentifierEnum::kChannelChange)
     {
         ChipLogProgress(Zcl, "IDENTIFY_EFFECT_IDENTIFIER_CHANNEL_CHANGE - Not supported, use effect varriant %d",
-                        identify->mEffectVariant);
+                        (uint8_t) identify->mEffectVariant);
         sIdentifyEffect = identify->mEffectVariant;
     }
